Add Navbar and css to use throughout site
