+++ frontend/src/App.js
@@ -1,14 +1,15 @@
+import './assets/stylesheets/reset.css';
 import React, { Component } from 'react';
 import {Route} from 'react-router-dom';
-import Greeting from './components/greeting/greeting_container';
+import NavBar from './components/navbar/navbar_container';
 import MainPage from './components/main_page/main_page';
 
 class App extends Component {
   render() {
     return (
       <div className="App">
-      <Greeting />
-      <Route exact path='/' component={MainPage}/>
+        <NavBar />
+        <Route exact path='/' component={MainPage}/>
       </div>
     );
   }
