--- frontend/src/components/greeting/greeting.js
@@ -1,15 +1,0 @@
-import React from 'react';
-
-class Greeting extends React.Component {
-    constructor(props) {
-        super(props)
-    }
-
-    render() {
-        return (
-            <h1>GO!</h1>
-        )
-    }
-}
-
-export default Greeting;
